+++ CHANGELOG.md
@@ -7,10 +7,10 @@ This section contains changes that have been committed but not yet released.
 ### Added
 
 - Enabled support for Nylas API v2.2
-- Add `metadata` field in the Event model to support new event metadata
-- Add support for filtering `metadata` using `metadata_key`, `metadata_value`, and `metadata_pair`]
+- Add Event Metadata support
 - Add support for new `RoomResource` fields
 - Add missing getters for Event.Recurrence fields
+- Add support for Neural API Sentiment Analysis, OCR, Signature Extraction, and Clean Conversations
 - Add getters for Time.timezone, Timespan.start_timezone, Timespan.end_timezone
 
 ### Changed